+++ src/main/java/org/opensearch/knn/quantization/models/quantizationParams/ScalarQuantizationParams.java
@@ -9,7 +9,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-
 import lombok.NoArgsConstructor;
 import org.opensearch.Version;
 import org.opensearch.core.common.io.stream.StreamInput;
@@ -26,12 +25,14 @@ import java.io.IOException;
 @Getter
 @EqualsAndHashCode
 @AllArgsConstructor
-@NoArgsConstructor
+@NoArgsConstructor(force = true)
 @Builder
 public class ScalarQuantizationParams implements QuantizationParams {
-    private ScalarQuantizationType sqType;
+    private final ScalarQuantizationType sqType;
     @Builder.Default
     private final boolean enableRandomRotation = QFrameBitEncoder.DEFAULT_ENABLE_RANDOM_ROTATION;
+    @Builder.Default
+    private final boolean enableADC = QFrameBitEncoder.DEFAULT_ENABLE_ADC;
 
     /**
      * Static method to generate type identifier based on ScalarQuantizationType.
@@ -65,6 +66,7 @@ public class ScalarQuantizationParams implements QuantizationParams {
     public void writeTo(StreamOutput out) throws IOException {
         out.writeVInt(sqType.getId());
         out.writeBoolean(enableRandomRotation);
+        out.writeBoolean(enableADC);
     }
 
     /**
@@ -74,13 +76,16 @@ public class ScalarQuantizationParams implements QuantizationParams {
      * @param in the input stream to read the object from.
      * @throws IOException if an I/O error occurs.
      */
+    // Constructor for deserialization
     public ScalarQuantizationParams(StreamInput in, int version) throws IOException {
         int typeId = in.readVInt();
         this.sqType = ScalarQuantizationType.fromId(typeId);
-        if (Version.fromId(version).onOrAfter(Version.V_3_1_0)) {
+        if (Version.fromId(version).onOrAfter(Version.V_3_2_0)) {
             this.enableRandomRotation = in.readBoolean();
+            this.enableADC = in.readBoolean();
         } else {
             this.enableRandomRotation = QFrameBitEncoder.DEFAULT_ENABLE_RANDOM_ROTATION;
+            this.enableADC = QFrameBitEncoder.DEFAULT_ENABLE_ADC;
         }
     }